requirements.txt
requests == 2.22.0
pandas == 0.25.1
sodapy == 1.5.4
pylint == 2.3.1
paramiko == 2.6.0
pyyaml == 5.1.2
